(PUP-5898) Make :undef be inferred as Undef (aka. bring back the undead)
Before this, if :undef was found in data coming from 3.x API into 4.x it
was inferred as Runtime[ruby, 'Symbol'] causing type mismatches to
occur.

This is now changed so that it is inferrd to be an Undef.

Unfortunately we had to revert the change that made :undef be treated as
an alien Ruby Symbol. It was intended to catch :undef leaking, but all
we achived was to catch ourselves in the act - and it is difficult to do
anything about this as some resurces do need 3.x call semantics and
others don't. It is not known a priori which ones.