(PUP-5899) Fix bug when handling empty yaml for hiera data provider
Before this commit, an empty yaml file would cause an internal error.

An empty yaml file is valid. When it's parsed it produces a `nil`. This
commit ensures that the yaml-based data provider for Hiera treats that
the same way as an empty hash.

For completeness, this commit also adds a test of what happens when
an empty JSON file is encountered. That should (and does) raise a
lookup exception indicating a parse error.