(PUP-5118) Manage duplicate file resources
Prior to this commit, the compiler could not handle duplicate file
resources when recursively in-lining metadata. This meant that a case
where a file is explicitly managed and also covered by its parent
directory which has recurse turned on would raise an error.

Now this case is handled by using remove_less_specific_files to
prune out duplicate resources. We will always keep the explicitly
managed version of the resource over the one generated by recursion.