(PUP-5873) Contain generated resources when parent has :in edges
If a resource in the manifest has a relationship to the parent, then
create an opposing relationship from each generated child to the
resource. If the original relationship is a `require`, then create a
`before`. Otherwise, if it's a `subscribe`, create a `notify`.

This handles cases where a single resource has a relationship to the
parent, where multiple resources have relationships, or a resource has
multiple relationships, only one of which needs to be to the parent.

This only handles require/subscribe, but not consume, and is missing
tests for adding notify relationships from each generated child to
resources that subscribe to the parent.