(PUP-5873) Memoize using the recursive directory's path if specified
Previously, we assumed the recursive directory's title was the
path. If it wasn't then we memoized the resource with the wrong key,
causing failures later when we tried to add a 'before' relationship
from the directory to any of its immediate children.

This commit ensures we memoize the resource using its path or title,
in that order. Note we don't have the same issue with generated child
resources, because we created them using the path as the title.