(PUP-5964) Enhance type asserter with deferred subject formatting
The TypeAsserter methods takes a `subject` parameter that explains in
what context the assertion takes place. It is often desirable to create
this subject string from things in that context (i.e. format a string).
Formatting strings that may or may not be used is however not a good
practice since the formatting itself is resource consuming.

This commit makes it possible to defer the formatting of the subject
string until an error actually has been detected.

This commit enables deferred formatting in one of two ways:

1. As an alternative to passing the subject as a String, it can now also
   be passed as an Array. When an Array is used, the first element is
   treated as a format string and the remaining elements are passed as
   arguments to that format string.
2. A block can be given provided to the assert call. The result of
   calling that block with the given subject as an argument will become
   the new subject. If the result is an Array, then it will be formatted
   as described in #1.