(maint) Properly stub catalog methods to prevent errors in tests
When running Puppet::Configurer::run_internal in spec tests,
attempting to send the report and write the class file error out
due to arbitrary files not existing on the filesystem. Stub
the methods responsible for these calls out so the error we are
trying to actually test can more cleanly be tested.