(maint) Fix pending spec test
The spec test was written so that the file resource's title (`c:/foo`)
and the expected resources title (`/foo`) would not match. This caused
the test to accidentally fail on Windows, where `/foo` expands to
`C:/foo`.

Fix that mistake, and skip the test on Windows, as it's specific to a
Linux master.