(PUP-5118) Update spec tests
Prior to this commit, there were no spec tests for duplicate file
resources with regards in-lining metadata. Add a test, but leave it
pending because some of our stubs are stepping on each other and the
tests must be refactored when we removed the to_ral call anyways.

Remove a duplicate test from the file type spec.