(PUP-5815) Handle cached catalogs with strict environment mode setting
This commit ensures that if a cached catalog is explicitly requested
and strict environment mode is also requested, we do *not* switch the
local environment to match the cached catalog, as we do in all other
cases. This will cause the run to fail if the cached catalog's
environment does not match the agent specified environment.