(PUP-5925) Treat undef as a valid value from a binding
Before this commit, a binding for a host class definition parameter
that was declared as `undef` was treated as "not found" and hence,
would yield an error if the parameter didn't receive a value some
other way.

This commit changes this so that an `undef` value found in a binding
is considered a valid parameter default. An `undef` value will however
never replace a value produced by a parameter default expression.