(PUP-5114) Implement retrieving files via content_uri
Inline metadata into the catalog. The file type uses that metadata when
available in place of making remote metadata requests.

This adds two new fields to the catalog: `metadata` and `metadatas`.
Metadata is a hash of File resource titles to the metadata for that
resource (for non-recursive resources). Metadatas is a hash of File
resource titles to a hash mapping recursive file sources to lists of
metadata found via recursion.

Obviates adding `content_uri` to the file type; it is now part of
`FileServing::Metadata`.

This approach was chosen instead of previous efforts because we found
much of the file type's logic was being duplicated in the master,
breaking type separation of concerns. Some knowledge of how files can be
queried is still needed to retrieve metadata, but much less than
previous work.

Major credits to @branan, @whopper, @joshcooper, and @HAIL9000 for
contributions to this PR and understanding the file type.