(PUP-5114) Make metadata.checksum_type authoritative in file type
The agent will always fail insync if the metadata checksum_type and the
resource's checksum parameter don't match. There are two instances
where this can happen: an HTTP/HTTPS request returning metadata that
was not coded into the resource, and the static catalog choosing a
different checksum_type than `digest_algorithm` on the agent.

When using static catalogs, the server will calculate metadata based on
the agent's `supported_checksum_types` setting. The checksum type can
differ from the agent's `digest_algorithm`, which is the usual default
for the file type.

Use the checksum_type returned by metadata. One caveat to this approach
is that the resource is no longer authoritative on checksum type. If
the metadata returns a checksum type, that will be used even if it
doesn't match an explicitly requested checksum type by the resource.
This is no different than other parameters/properties that are updated
from source like whether we're creating a link/file/directory.

Note that the file_checksum schema requires that metadata.checksum_type
is always set.