(PUP-4748) Allow purging of resources that have dependencies
Due to how resource generation was implemented, the purge check for
resource dependencies was actually broken until PUP-1963, which
accidentally caused that check to work.

Since that check has been broken for a long time, and users generally
have an expectation of being able to attach dependencies to a purge,
the check should be removed.