(PUP-5889) Add a --strict=[ignored, warning, error] setting
This adds the --strict setting to puppet. It can be set to
'ignored', 'warning' or 'error'. The default value is 'ignored'.

This commit only adds the setting itself and a unit test that the
expected values can be set, and that a descriptive error is raised if
not.

This also adds a generic :symbolic_enum settings type to avoid having to
compare against string values at runtime.