(PUP-5121) Move inlining into compile profile block
Previously the inlining existed as a separate step which was profiled
and benchmarked independently of the base compile. This made sanely
producing comparisons between static compiles and non-static compiles
challenging, so this commit moves the inline_metadata call into the base
compile benchmark block. This has the benefit of still allowing the
inline_metadata call to be independently benchmarked, while still
aggregating it in the default catalog compilation metrics.