(PUP-5963) Restore functionality of recursive tidy resources
Recursive tidy resources depended on the resources returned by tidy's
`generate` to be handled by the additional_resource_generator in a
specific order. This restores that order. I actually think that this
is "correct" for all depth-first resources, and any depth-first
generation has been wrong since PUP-1963. This also adds a test for
Tidy, since we apparently didn't have any.