(PUP-5964) Make TypAsserter use the TypeMismatchDescriber
Before this commit, TypeAsserter had it's own poor-mans solution for
how to format the string used for reporting a mismatch. This changes
in this commit so that the TypeMismatchDescriber is utilized instead.

The commit also changes a couple of places where the TypeAsserter
is used since the generated messages no longer made sense.