(PUP-5964) Improve type mismatch describers ability to generalize
Before this commit, the TypeMismatchDescriber would sometimes
generalize too much because the classes of the expected and actual
types didn't match (and PStructType != PHashType) or not generalize
although it was possible (generalized form of actual type still not
assignable).

This commit improves the logic that decides where or not to generalize
the types before reporting the mismatch.