(PUP-5959) Derive static catalog metadata lookup defaults from File type
Previously, while inlining file metadata for the static catalog
our default fallbacks for the :links and :source_permissions parameters
were hardcoded to match the file type.

This commit removes those hardcoded values in favor of querying a
dummy file resource for its actual defaults. An instantiated
resource is necessary due to Puppet::Type parameter defaults not being
resolvable until a specific provider is known. Since :links and
:source_permissions are provider-independent, it is safe to instantiate
the resource on the master.