(maint) Use new tmpfile for each agent in direct puppet test
Previously, the static catalog drift remediation test created
the agent test file path using `agent.tmpdir`, which is really
an alias to `agent[0].tmpdir`. This would normally be fine,
except in the case where the agents array contains both
Linux and Windows hosts, as the temporary directories between
the two differ.

Instead, generate a new tmpfile for each agent in the agent array
and update the test file resource on the master accordingly.