(PUP-5768) Fix problematic usage of ENV for Windows
The ENV class in Ruby is known to corrupt environment variables which
contain UTF8 characters in their names and values.  This commit adds an
operating system safe way of accessing environment variables as part of
the Puppet::Util module.  The files in this commit have been ammended to
use the safe invocation module instead of the unsafe ENV class.

This commit also updates the #which method in Puppet::Util to reduce the
number of calls to get the PATHEXT environment variable. Previously PATHEXT
is evaluated for every element in the PATH wheras this change will evaluate
PATHEXT once.  This change will only affect Windows Operating Systems