(PUP-5903) Refactor E4ParserAdapter to give access to cached parser
Before this it was difficult to get the parser instance used by the
compiler.

Now, it is possible to get the shared EvaluatingParser instance via a
class method. (Later it is of value to make additional refactoring in
this area as there are several places where EvaluatingParser is cached
and handled.

It is bad to create new instances of the evaluator as its state is of
value (it builds up an index of polymorphic dispatch entries).