(PUP-5986) Modify TypeParser to allow both Scope and Loader as context
Before this, the TypeParser found the loader to use when resolving type
references via a given scope. This is difficult to use in situations
where an applicable scope is not available, but where the loader is
known.

This modifies the TypeParser to accept a "context" which can be a Scope
or a Loader.