(PUP-5866) Per-environmment control of static catalogs
This test verifies that the environment.conf setting:
static_catalogs = [true | false] controls whether catalogs for that
environment are static catalogs or not.  The same setting in
puppet.conf will determine the default behavior, that it if the
setting is not specified in environment.conf.  If the setting is not
specified in either configuration file, then static catalogs are
enabled by default.

[skip ci]