(PUP-5959) Add test to ensure static catalog file params match defaults
When inlining file metadata, we sometimes hardcode the default values
for the `links` and `source_permissions` parameters to match
the real file type defaults.

This commit adds a new test case to ensure that the values we've
hardcoded actually match those defaults. Should they ever change,
this test will fail, thus notifying us that our hardcoded values
need to change.

This is being done with a test rather than retrieving the defaults
during compilation due to the fact that determining said defaults
requires instantiated file resources, which depend on knowing which
file provider to use. Creating a new dummy file resource on the
master is clunky and inelegant.