(PUP-5991) Add failing test for <drive>:\DEV on Windows
Although C:\Dev\.... is a valid path on Windows, some other code and test
may regard it as a path to be ignored. e.g. /dev/null resolves to
C:\dev\null on Windows.  The unit test fails are ambiguous (e.g. expected
nil) and do not convey the underlying issue.  This commit adds a test that
fails if the working directory has a root directory of 'dev'