(maint) Correct error in the Puppet lookup command test.
Before this fix, this test fails when there is more than
one agent.  The beaker variable, agent is treated was treated as
a scalar.  When there is more than one agent, the beaker variable
is an array.  The fix copies the agent to a scalar variable, checking
if it is an array, in which case the scalar variable takes the value
of one element of the array.