(PUP-5995) Change SourcePosAdapter to delegate extraction to locator
Before this, the SourcePosAdapter assumed that the source string would
be sliceable based on the offset and length reported by the Locator.
This was wrong as the locator may use byte or Char representation
depending on Ruby platform (there are two kinds of Locators). So far,
all rubies measured perform better when recording byte offsets rather
than Char offsets when using frozen strings.

The problem meant that the SourcePosAdapter methods 'extract_text' and 
'extract_tree_text' would extract source that was offset one r more
bytes for every preceding multi-byte character. This problem was visible
in Puppet Strings, and in the ClassInfoService where the source for a
parameter default value expression was extracted.

This is now changed so that the Locator is responsible for performing
text extraction. The SourcePosAdapter now delegates the task of
extracting the text so it gets the correct (byte or char)
implementation.

Note that only the byte based implementation is currently used and there
are no tests for the character based implementation.