(PUP-5099) Add code_id to environment catalog from request
This follows the model of agent catalog compilation, passing code_id
from the HTTP request to the environment compiler, which adds it to the
catalog. We then read code_id from the environment catalog and include
it in our response.

This allows environment catalogs and agent catalogs to be correlated
with one another and against a specific version of code.