(PUP-6009) Fix typos in split function documentation
The documentation for split shows that a Regexp type can be used, but
uses ( ) insted of [ ] for the type arguments.

When changing this, the 4.x version of the function was found to have
garbage at the end of the documentation (copy paste error from the 3.x
versions).

The 3.x version of the documentation did not show the Regexp type
example.