(PUP-5902) Make Type stringifier use white space more consistently
The type stringifier consistently adds a white space after each comma
but it does not use white space when it outputs the '=>' operator. This
commit ensures that the '=>' is surrounded by whitespace.

Later, we might want to consider using different modes. One that is
compact (no whitespace anywhere), one semi-compact (what this commit
provides) and one that uses pretty formatting with newlines and indents.