(maint) Support multi-agent acceptance configurations
Previously the test assumed there was a single agent, but when run with
multiple agents, beaker's `agent` method returns an array. And since
`on` accepts an array of hosts, and returns an array of results, we were
trying to call Array#stdout and failing.

This commit updates the test to use the first non-master agent. The test
then runs `puppet lookup --node <agent>` on the master. Since the test
is exercising lookup functionality on the master, we don't need to
repeat it for every agent.

The test also failed on master platforms that didn't have system ruby
installed in /bin/ruby, e.g. debian8. Instead we use privatebindir to
resolve ruby, which works for aio and git.

[skip ci]