(PUP-5902) Output type alias mismatches with aliased type
Before this commit, a type alias would simply be output as the name of
the alias in a type mismatch. This is now changed so that an alias
is output as A=T where A is the name of the alias and T is the type
that it represents. The change applies is also applied to aliases
nested inside of other types.