(PUP-5902) Move stringifying of Puppet types to a separate class
Before this commit, the TypeCalculator was responsible for creating a
string from a type in the Puppet type system. This commit moves this
responsibility to a separate TypeStringifier class.

The commit also changes the logic so that formatting of arrays and
hashes are moved to a common place. The intention is to make it easier
to add different types of formatting such as 'compact' or 'pretty'.

A minor bug in 'debug_string' is also fixed in this transition. The
old implementation would only output strings at the top level in debug
format while types nested in other types would still be output using the
normal 'string' method and hence without debug info. Now the debug
format is retained when stringifying contained types.