(PUP-5901) Refactor common type behavior into abstract base class
The PType, PNotUndefType, PIterableType, PIteratorType and
POptionalType all contains exactly one type and therefore share a lot
of behavior such as accept, initialize, hash, ==, and generalize. This
commit moves these methods to a common abstract super class.

The commit also improves the logic of the generalize methods so that
they return self when the generalization changes nothing.