(PUP-5901) Add normalization capability to the Puppet Type system
This commit adds the logic necessary to normalize types such as
Optional, NotUndef, and Variant. The normalization does not change
the type characteristics but it removes redundant constructs such as
an Optional contained in a NotUndef (or vice versa).

The normalization will also merge Enums contained in a Variant into
one Enum (and possibly drop the Variant). The same is true for Patterns.
Overlapping ranges are merged and in case of Integer ranges, adjacent
ranges are also merged.

A Variant that contains an Undef is transformed into an Optional with
normalized contents.