(PUP-6024) Restrict sources of inlined file metadata
Previously, we would inline file metadata for any file in the
per-environment directory, however, puppetserver would only serve file
content for files within the per-environment modules/<module>/files directory.
As a result, we were creating a catalog whose static file content could
not be retrieved in some situations, notably when using a control repo
with a source like:

    puppet:///modules/profile/puppetmaster/update-classes.sh

which uses a custom modulepath in environment.conf to serve the
file from:

    site/profile/files/puppetmaster/update-classes.sh

Note the module 'name' in this case is 'profile'.

This commit ensure we only inline file metadata when using the 'modules'
mount and the resolved path is in a per-environment directory with a
relative path of the form: <base>/<module>/files/