(PUP-6033) Normalize source parameter when inlining metadata
Previously, if a source parameter had a trailing slash(es), we would
inline the metadata with the trailing slash. However, the agent always
strips trailing slashes from its sources, and it was not able to lookup
the inlined recursive metadata from the catalog.

This commit normalizes the source parameter when inlining file metadata,
so that `catalog.recursive_metadata[title][source]` works as expected.

Strangely, puppet strips trailing slashes for all file resources, not
just directories, so we apply the same logic when inlining
non-recursive metadata too.