(PUP-6036) Escape path when validating content_uri
The content_uri metadata is composed from the source scheme, host, and
port, and the path of the file on disk, relative to the per-environment
directory. Since the path can contain characters that must be
percent-encoded, e.g. 'foo%bar', we need to escape the path prior to
trying to parse the URI.