(maint) Unwrap arrayification of config_version from execute call
Previously, the config_version was wrapped in an array for the
Puppet::Util::Execution.execute call because an older instance of the
method (when it was Puppet::Util.execute), would raise an exception if
the argument to it was not an array. The current implementation accepts
both string and array, so the wrapping of it in an array is no longer
required.