(PUP-3018) Sanitize resource lists with nested empty arrays
Before this if the meta parameters for resource ordering ended
up being set to a nested structure of empty arrays, the resource could
not be converted to RAL form.

This is now handled by the Parser::Resource class when it is converted
to a hash. (This is the API that is used between "plain Resource" and
"parser Resource") when the catalog is "reconstructed" to a RAL catalog.