PUP-6048: always create a basic default environment
At the moment puppet only creates a basic default environment, if the
environment is named production. I don't see a reason why production needs a
special handling in this case. Also changing this behaviour will resolve the
problem, that 'puppet facts --environment test' throws an error if the
environment doesn't exist in the file system.