(PUP-5965) Add the 'new' function and make it callable via T()
This adds a 'new' function, and the ability to call this directly
from a type on the form `T()`.

This commit adds:
* the new 4.x function
* a new function in the Integer type
* a new function in the Optional type
* updated checker (to allow calling a type)
* updated evaluation to call the function named new when the value
  is a type.

The design decision was to use a Puppet function and that types
can return a new Function class when asked for its new_function.

This was chosen so that later, it is possible to get the type
signature of the new_function for a type.

This is a series of commits where this first commit only support
Integer, and Optional[Integer] as well as general error handling
for the Any type.