(maint) Fix problem with type aliases being written into parent
Before this, there was a problem with type aliases as they ended up
being loaded into the parent loader, instead of the local loader just
created for aliases.

This means that all functions in the same module ended up sharing the
context in which functions are loaded. This can lead to errors if they
happen to use the same alias.