(maint) Improve error when function local type alias is bad
Before this, a typo in a function local type alias would lead to a
NotImplemented (find) error. This is obviously bad as you do not see the
actual typo.

This commit adds a PredefinedLoader that is intended to be used when a
loader is pre-populate with loaded items before it is used.

This loader also requires that type references for Resources are written
out in long form as it will error if an attempt is made to load a non
existing type. The normal behavior is otherwise to return nil when
something is not found. That does not work well in the use cases
intended for the PredefinedLoader as it means that a legit call will be
considered to not match arguments. When that is reported it hides the
fact that there is a typo.