(maint) Remove ancient broken test involving code removed > 1 year ago
This removes testing that the "puppet_backend" for hiera (in the
puppet code base) can be loaded if configured. The result is that it
tries to load a file that was deleted over a year ago.

The reason why this problem occurs now, is because of order dependencies
among the tests. When order depending tests were fixed the now removed
test became order dependent. When it works it got an already configured
hiera, and the test assumed everything worked, since the only thing the
test did was no ensure it was possible to load a configuration
containing "puppet_backend".

When the puppet_backend freak was removed, this test became void, but
because of ordering issues it has not been seen until now.

This commit removes the broken test and the helper setup method
that configured the "puppet_backend".