(PUP-6070) Ensure that hash numbers don't exceed 64-bit limit
Before this commit, the hash computation of deeply nested structures
of objects could result in a hash value so large that it exceeded
what fits into a 64-bit integer. This lead to problems with JRuby
since it will error out if such objects are added to an array that
is then later asked to produce a hash. The error is:
```
RangeError: bignum too big to convert into `long'
```
This commit ensures that no such numbers are produced by replacing all
multiplications and additions in the #hash methods with logic that
either does an XOR of the values or, when several values are involved,
puts them into an array that is then asked to produce the hash.