(PUP-6075) Ensure known resource type registration use downcased names
Before this commit, the logic for registering known resource types in
the static loader was `TypedName`instances with capitalized names. This
resulted in that the `TypeParser` could never find these types since it
always passes a downcased name to the loader.

This commit changes the registration to instead use a downcased name
(for the registration, not for the registered resource). It also creates
a no-op '#resolve' method on the `PAnyType` so that type `TypeParser`
doesn't need to check what kind of type it is that the loader returns
before calling that method.