(PUP-6028) Clean up use of app-management in unit tests
When app-management was first introduced, switching from and to an
app-management enabled runtime was a complex task. Conceptually it was
also believed to mainly affect the parser tests due to new keywords and
consequently, the method `with_appmanagement` was added to the pops
parser_rspec_helper.rb file.

None of this is true today. Switching is as simpel as setting the
Puppet[:app_management] and it will now also affect the loader since
new resource types are introduced. This commit therefore removes the
no longer needed method from the parser_rspec_helper.rb and makes the
necessary changes to the tests.