(PUP-6028) Let PResourceType retain capital letters in the type name
Before this commit, a PResourceType would always use a downcased name.
This makes type mismatch messages harder to read because a resource type
is often presented by it's name only.

This commit changes this so that the original name is retained in the
type. The hash and eql? methods will however still use a downcased
name to ensure that a comparison is always case insensitive.