(PUP-6028) Prevent resource creation in Catalog#resource method.
Before this commit, the Catalog#resource method would always create a
new resource in order to create a key to lookup an existing resource.
It would also create two keys (title key and uniqueness key) even if
the existing resource was found using the first key.

This commit changes the behavior so that a title key can be created
without first creating a resource. If a resource is then found with
that key, no further processing is needed. This happens in the majority
of cases.

The change shows a 5% performance increase on the `many_modules`
benchmark.