(PUP-6028) Add loader that loads 3.x resource types
Before this commit, a 3.x resource type could not be told apart from
an unresolved reference unless the resource type was statically loaded.
This commit adds a new {{Runtime3TypeLoader}} that will delegate the
lookup to the {{TypeCollection}} `known_resource_types` in the current
environment. This instance will in turn delegate to the 3x loader in
case a resource is not known.