(PUP-5482) Clear missed types after catalog compile
When caching misses on loading types, if a new type is used in a
manifest before it is available as a type, even if that type is later
added, compilations will continue to fail for the life of the server
instance. This commit avoids that problem by clearing the type misses
after the catalog compile has completed.