(maint) Fix bug when comparing Hash type to a Struct
This commit fixes a bug PStructType#_assignable? that was triggered
when passing a PHashType with correct key and element types but an
undefined size type (the error was triggered from a unit test and
probably not very common).

A hash instance is normally inferred to a Struct or, in some cases, a
Hash with a defined size.