(PUP-5844) Add Object type to the Puppet Type system
This commit adds a new Object type to the Puppet Type system. An
Object may have attributes, functions, equality, checks, and
annotations declared. It may also inherit from a parent type.

The Object type can be seen as the interface of an implementation, in
that it describes the contract that the implementation must fulfill.
The type will however not contain the actual implementation logic.

Things not included in this commit:
1. Logic around 'checks' (awaits spec)
2. Logic around 'annotations' (awaits spec)
3. The new-operator. Will be implemented together with 'checks' logic