(PUP-6128) Retain case in QualifiedReference
Before this commit, all qualified references were downcased immediately
in the AST model. Attempts where then made to restore the case when
reporting errors by capitalizing all segments in the name. So names
like `MyGreateResource` became `Mygreatresource`.

This commit changes the QualifiedReference so that it instead stores
the original name and will deliver the downcased name when asked to
do so. The logic is also changed in several places to avoid repeated
downcasing and capitalizing.