(PUP-6105) Avoid comparing symbol to string in terminus
Before this, there was an attempt to create a symbol from an empty
string (or worse case from nil). With the removal of monkey patching
symbol to compare equal to string and introducing warnings, it is
important to clean up odd usage of string symbol comparisons.

This is now changed to deal with the error condition of nil/empty string
before converting to symbol and using the value.

This also makes a small refactoring to the variable name as the
implementation in general uses a very confusing style where some
attributes have readers but no writers this resulting in confusion over
if the use of the attribute is a local variable or a function call to
the attribute reader.