(PUP-6139) Log HTTP errors at debug level
Previously when an indirection raised an HTTPError the error was logged
at the info level. This causes a lot of noise for resources such as a
file resource with sourceselect, where several sources may 404 while
trying to find a successful source. This commit moves the message to the
debug level to avoid that problem.