(PUP-6142) Only warn once for missing module dependencies
This commit changes the warning for missing module dependencies
to be less verbose.  It also adds support for the `:strict`
setting.  If `:strict` is set to `:warning`, then we only log
the missing module dependency once (per ruby instance).  If `:strict`
is set to `:error`, then we raise a LoaderError.  If `:strict`
is set to `:off`, then we don't log anything.