(PUP-6140) Acceptance test for init.d services
An acceptance test for managing init.d services, particularly on
systems where the systemd provider is the default. This doesn't work
without PUP-2744, which ensures the fallback on RedHat systems is the
RedHat provider - which works - instead of the init provider - which
doesn't on RedHat.

It also helps verify that managing SysV init scripts continues to work
after a fix for PUP-5296/PUP-5825, where enabled status was reported
incorrectly.

[skip ci]