(maint) Refactor TypeParser to use a hash type name => type
Before this, the information that maps string name to type instance for
build in types was expressed only in a case expression for evaluation of
a QualifiedReference. This used to call the TypeFactory to get a default
instance for each looked up name. Now, instead of using a case switch to
get to a call to TypeFactory, which then returns a singleton default
type, this is directly stored in a hash map. This removes one level of
indirection and provides the ability to get this map in other parts of
the system.