(PUP-1526) Add environment settings for package type
This patch adds to the package type the `:settable_environment` feature
and the `:environment` parameter used by that feature.  `:environment`
is a hash of key=>value pairs; keys must be non-empty strings, and
values must either be `undef` or non-empty strings.  `undef` values are
translated to ruby `nil` values before being passed to the provider.

The `yum` package provider is updated to provide the
`:settable_environment` feature.

Finally, the `spec` tests for the `package` type and `yum` provider
are updated to test for the `:settable_environment` feature and the
`environment` parameter.